pyproject.toml
[tool.poetry]
name = "rockscissorspaper"
version = "0.1.0"
description = ""
authors = ["Alexander-Zaychenko <alexzay2012@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"
telebot = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
